Add a preventDefault param to AElements that controls href behavior
## Background

As of 1da0b382, all AElements no longer lead to page loads, but window.url changes.
This is generally consistent with the SPA paradigm that Kweb leans towards.

However, this is a very breaking change for anyone using AElements. Overnight, folks
updating Kweb versions had all their hrefs changed to not do full page loads. This is
both unexpected, as the doc was kinda buried in the AElement code (rather than the class
KDoc, or the aelement helper Kdoc), but also just plain buggy - e.g what if you serve
other things using Ktor from the same domain?

Until now, there was no way to control this behavior, even though in the end it's just
a boolean toggle in the source, since it's not exposed. Users had to create their own
implementation of AElement with the desired functionality.

## Change Description

This changeset:
* Adds a `preventDefault` parameter to the `AElement` and the `a` DSL helper, defaulting to true
* Uses the `preventDefault` param to control the href behavior of pageload/no-pageload
* Updates the docs and exposes this behavior both on the `AElement` and the `a` DSL helper

All in all, merging this will keep the current functionality as is for everyone, but it will
allow configuring per element for users who need that granularity.